+++ Target/Function/Build.d.ts
@@ -81,6 +81,6 @@ import type Interface from "../Interface/Build.js";
 declare const _default: Interface;
 export default _default;
 export declare const Exec: import("../Interface/Exec.js").default;
-export declare const Merge: <Target extends object, Ts extends readonly unknown[]>(target: Target, ...objects: Ts) => void;
+export declare const Merge: <Target extends object, Ts extends ReadonlyArray<unknown>>(target: Target, ...objects: Ts) => void;
 export declare const Pipe: string[];
-export declare const extname: typeof import("node:path").extname, basename: typeof import("node:path").basename, relative: typeof import("node:path").relative, dirname: typeof import("node:path").dirname, normalize: typeof import("node:path").normalize, sep: "\\" | "/";
+export declare const extname: typeof import("node:path").extname, basename: typeof import("node:path").basename, relative: typeof import("node:path").relative, dirname: typeof import("node:path").dirname, normalize: typeof import("node:path").normalize, sep: "/" | "\\";