+++ Target/Function/Exclude.d.ts
@@ -1,3 +1,4 @@
+export default _default;
 /**
  * @module Exclude
  *
@@ -11,7 +12,6 @@
  * @returns True if the Path matches any pattern in From, false otherwise.
  *
  */
-declare const _default: (Path: string, From: string[]) => boolean;
-export default _default;
-export declare const posix: typeof import("node:path"), sep: "\\" | "/";
+declare function _default(Path: string, From: string[]): boolean;
+export declare const posix: typeof import("node:path"), sep: "/" | "\\";
 export declare const _Regex: (Glob: string) => RegExp;