+++ Target/Variable/ESBuild.d.ts
@@ -39,4 +39,4 @@ export declare const On: boolean;
  */
 declare const _default: BuildOptions;
 export default _default;
-export declare const sep: "\\" | "/", posix: typeof import("node:path");
+export declare const sep: "/" | "\\", posix: typeof import("node:path");